backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
requests
aiohttp
PyPDF2
python-docx
transformers
torch
sentence-transformers
numpy
scikit-learn
matplotlib
seaborn
aiofiles
cors
